Add time periods
Implement mechanism called *time periods*. Each resource could be
now combined with a period so that we get the data only from
that time period.

Each period consists of the **start** of the period and theaggregation
**window**. All the results inside one window are summarized into one
number.

Start of a period is defined relativelly from the cuurent time.
End of every period is the current time.

The default time period - used for currently precached data and
default views is **1y** (with aggreagtion window **1w**).