Add JS on button click for dynamic content reload
Names of all resources needed in a single page must be passed to the
template during rendering. Thanks to this each button knows which
resources to request when it is clicked.

On period/time button click two things happen:
    - Page URL is modified with corresponding period as URL param
    - For each resource name a new polling interval is set

During polling for resources the resource API is queried. When the
requested data finally arrives the corresponding interval is cleared
and the corresponding component is redrawed with the new data.

The corresponding component is identified by its `id` which is the same
as its `resource_name`. (This limits the number of components using a
the same resource to one. To overcome this `name` or `class` may be used
isted.)