Cache user specific data correctly
Introduce new query parametr called `token` which is
a hash of all user device tokens.

This new `token` will be used as a part of Redis key so that distinct
user data can be cached separately.

This new `token` will be also reused in resource GET http request
so that it can be localy cached correctly.