Use precise intervals in SQL queries
Timescale `time_bucket` function uses semi-static windows (aligned with
midnights / half and quarter hours) for data
grouping. This means that the first and/or last window may be shorter
than the others and thus may contain less data. This results in steep
increase or deacrease in the beginning or at the end of the data
graph.

In this commit, the queried interval is precisely set to match timescale
time windows so that the first one and the last one have the same size
like the others.

More about `time_bucket` function:
https://docs.timescale.com/api/latest/hyperfunctions/time_bucket/#time-bucket