Mark graph points with bucket (window) ending time
Before this commit, the points in graph (A,B,C,D,E,F) which display
number of incidents/attackers/passwords or port scans, were marked with
the starting (now changed to ending )time of their respective time window.

That means that point B, marked with 00:01:00, used to display 6 incidents that happened between one
minute after midnight (0:1) and 0:2. See the following example:
```
Incidents:
    - 0 in <00:00:00, 00:01:00)    .. displayed as point A
    - 6 in <00:01:00, 00:02:00)    .. displayed as point B
    - 0 in <00:02:00, 00:03:00)    .. displayed as point C
    - 0 in <00:03:00, 00:04:00)    .. displayed as point D
    - 6 in <00:04:00, 00:05:00)    .. displayed as point E
    - 6 in <00:05:00, 00:06:00)    .. displayed as point F

|             Example of old incidents graph                 |
|7                                                           |
|6       B              E----F                               |
|5      / \            /                                     |
|4     /   \          /                                      |
|2    /     \        /                                       |
|1   /       \      /                                        |
|0  A         C----D                                         |
|  0:0  0:1  0:2  0:3  0:4  0:5                              |
```

This was:
 - not intuitive as the curve was rising between A(0:0) and B(0:1) where
no incidents happened
 - default behaviour of timescale `time_bucket*` functions which simply
truncate time column so all the times are equal to the starting one

In this commit each time was increased by bucket(window) length so the
graph points are marked with the ending time of each bucket(window). See
the example which uses the same data:

```
Incidents:
    - 0 in <00:00:00, 00:01:00)    .. displayed as point A
    - 6 in <00:01:00, 00:02:00)    .. displayed as point B
    - 0 in <00:02:00, 00:03:00)    .. displayed as point C
    - 0 in <00:03:00, 00:04:00)    .. displayed as point D
    - 6 in <00:04:00, 00:05:00)    .. displayed as point E
    - 6 in <00:05:00, 00:06:00)    .. displayed as point F

|             Example of new incidents graph                 |
|7                                                           |
|6       B              E----F                               |
|5      / \            /                                     |
|4     /   \          /                                      |
|2    /     \        /                                       |
|1   /       \      /                                        |
|0  A         C----D                                         |
|  0:1  0:2  0:3  0:4  0:5  0:6                              |
```

This is more intuitive as the curve rises in windows where the incidents
happen and also because each point shows incidents that happend up to
that point (but excluding its exact time - the windows are closed from
left and open from right. E.g. incident that happened in 00:01:00.00000 is
included in point B).

More in general, the points display incidents that happened BETWEEN exact
times and therefore should be always displayed with precision.
E.g. displaying day-long windows only with 'day' precision would be
misleading as users would expect data from that 'day' whereas the data
would be from period ENDED BY that day (=the day before).