statistics: Increase resource cache TTL
Increase resource cache TTL for three periods:
    - 12 Hours (`12h`)
    - Day (`1d`)
    - Week (`1w`)

In case of `12h` and `1w`, the increase is just to make the time reserve
bigger. It is probably not needed.

In case of `1d`, it is a bug fix - the former cache TTL was 25 minutes,
which was 5 minutes less than refresh interval. It means that cache used
to expire few minutes before it was refreshed. In addition with
`start_delay_bonus` (3 mins and few seconds), the day period cache was
unavailable between every hour:25 - hour:33 and hour:55 - (hour+1):03.
As a result, the user was unable to see any precached data and some
query-specific data from that period in the respective moments.

This commit increases resource cache TTL of `1d` period to `50` minutes
so there is always enough time reserve even for unpredicted situations.