statistics: Fix refresh timeout for user resources
Cache TTL for user specific resources is lower so that user-specific
resources are not kept for too long. Although that the refresh timeouts
were the same like for other resources. As a result, the users were not
able to see their data after the short cache expired.

This commit reduces the refresh timeout for user-specific resources.
The USER_TTL might be increased in the future - in a situation when it is better
to store more data in cache than do more frequrnt DB queries.